* Addressing review comment - static functions to save us from using costly hash lookup operations * Also removed additional header * Renamed the container handler class
